# Performance Optimizations: Caching, Throttling, and Adaptive Performance Monitor