--- a/index.html
+++ b/index.html
@@ -82,6 +82,7 @@
 		<script src="./scripts/score/showScoreDots.js"></script>
 		<script src="./scripts/others/toast.js"></script>
 		<script src="./scripts/playMode/updateScoresTable.js"></script>
+		<script src="./scripts/cache/cacheManager.js"></script>
 	</head>
 
 	<body>