--- a/scripts/ball/edgeTouch.js
+++ b/scripts/ball/edgeTouch.js
@@ -1,6 +1,6 @@
 function isEdgeTouch(edge = ''){
-    const fieldCoords = getCoords('field', true);
-    const ballCoords = getCoords('ball', true);
+    const fieldCoords = getCoords('field', true, true);
+    const ballCoords = getCoords('ball', true, true);
 
     return +(fieldCoords[edge].toFixed(2)) === +(ballCoords[edge].toFixed(2));
 }