--- a/scripts/others/getCoords.js
+++ b/scripts/others/getCoords.js
@@ -1,5 +1,14 @@
+const CACHEABLE_KEYS = ['coords-field-false', 'coords-field-true'];
+
 function getRelativeToField(elemCoords){
-    const fieldCoords = getCoords('field');
+    const cacheKey = 'coords-field-false';
+    let fieldCoords;
+    
+    if (window.CacheManager.has(cacheKey)) {
+        fieldCoords = window.CacheManager.get(cacheKey);
+    } else {
+        fieldCoords = getCoords('field');
+    }
 
     return {
         left: elemCoords.left - fieldCoords.left,
@@ -10,8 +19,13 @@ function getRelativeToField(elemCoords){
 }
 
 function getCoords(id, relativeToField = false) {
-    const elem = document.getElementById(id);
+    const cacheKey = `coords-${id}-${relativeToField}`;
     
+    if (window.CacheManager.has(cacheKey)) {
+        return window.CacheManager.get(cacheKey);
+    }
+
+    const elem = document.getElementById(id);
     const elemBorderWidth = parseFloat(getComputedStyle(elem).borderWidth);
     const { left, top } = elem.getBoundingClientRect();
 
@@ -24,6 +38,10 @@ function getCoords(id, relativeToField = false) {
 
     if(relativeToField) {
         elemCoords = getRelativeToField(elemCoords);
+    }
+
+    if(CACHEABLE_KEYS.includes(cacheKey)) {
+        window.CacheManager.set(cacheKey, elemCoords);
     }
     return elemCoords;
 }