--- a/scripts/score/removeScoreDotIfEarned.js
+++ b/scripts/score/removeScoreDotIfEarned.js
@@ -1,14 +1,14 @@
 function checkIfScoreDotTouchBall(scoreDot, ball){
     const ballCenter = {
-        x: ball.left + ball.width / 2,
-        y: ball.top + ball.height / 2,
-        r: ball.width / 2,
+        x: ball.left + BALL_WIDTH / 2,
+        y: ball.top + BALL_HEIGHT / 2,
+        r: BALL_WIDTH / 2,
     }
 
     const scoreDotCenter = {
-        x: scoreDot.left + scoreDot.width / 2,
-        y: scoreDot.top + scoreDot.height / 2,
-        r: scoreDot.width / 2,
+        x: scoreDot.left + SCORE_POINT_DOT_WIDTH / 2,
+        y: scoreDot.top + SCORE_POINT_DOT_HEIGHT / 2,
+        r: scoreDot.radius,
     }
 
     const distance = Math.sqrt(
@@ -22,24 +22,20 @@ function checkIfScoreDotTouchBall(scoreDot, ball){
     return false;
 }
 
-function removeScoreDotIfEarned(){
-    const ballCoords = ball.getBoundingClientRect();
+function removeScoreDotIfEarned() {
+    const scorePointCoords = window.CacheManager.get(ALL_SCORE_POINT_DOT);
+    if (!scorePointCoords) return;
 
-    let scoreDots = window.StaticCacheManager.get(ALL_SCORE_POINT_DOT);
+    const ball = getCoords('ball', true);
+    
+    scorePointCoords.forEach((scorePoint) => {
+        if (scorePoint.isCollected) return;
 
-    if(!scoreDots){
-        scoreDots = document.querySelectorAll('.score-point-dot');
-        window.StaticCacheManager.set(ALL_SCORE_POINT_DOT, scoreDots);
-    }
-
-    scoreDots.forEach((scoreDot) => {
-        const coords =  scoreDot.getBoundingClientRect();
-
-        if(checkIfScoreDotTouchBall(coords, ballCoords)){
-            scoreDot.remove();
+        if (checkIfScoreDotTouchBall(scorePoint, ball)) {
+            scorePoint.element.remove();
+            scorePoint.isCollected = true;
             playSound('coin-collect-sound');
-
             updateCurrentScore(++currentScore);
         }
-    })
+    });
 }