## Overview
Resolves #36 [On Slower Devices, game becmoes laggy, We should try to improve the performance of the game].

This PR implements several performance optimizations resulting in significant improvements to game performance across different devices:
- Coordinate caching system for field elements
- Throttled trajectory calculations
- Adaptive performance monitoring system

## Performance Improvements 📈

### Field Coordinate Access
- **Without Caching**: 9.47ms (avg)
- **With Caching**: 3.75ms (avg)
- **Improvement**: 60.42% faster ✅

### Projectile Motion Performance
- **Base Performance**: 7,997.62ms (avg)
- **With Caching**: 5,844.92ms (avg)
- **With Caching + Throttling**: 5,390.76ms (avg)
- **Total Improvement**: 32.6% faster ✅

## Technical Implementation Details 🛠️

### 1. Coordinate Caching System
- Implemented Map-based caching for field coordinates
- Cache invalidation on window resize
- Selective caching for frequently accessed elements (ball, field)
- Cached keys: `coords-field-false`, `coords-field-true`

### 2. Trajectory Optimization
- Implemented throttling (20ms) for trajectory calculations
- Reduced unnecessary recalculations
- Optimized `getCoords` function with selective caching

### 3. Performance Monitor
- Added new `performanceMonitor.js`
- Automatic performance detection system
- Smart adaptations for low-performance devices:
  - Simplified trajectory visualization
  - Reduced animation complexity
  - Optimized CPU-intensive operations

## Benefits 🎯
- Improved frame rates
- Reduced CPU usage
- Better responsiveness
- Enhanced cross-device compatibility
- Smoother gameplay experience

## Testing Notes 🧪
Performance tests were conducted with:
- Multiple runs for consistency
- Various device configurations
- Different game scenarios (vertical movement, projectile motion)

## Future Considerations 🔄
- Further optimization opportunities in animation systems
- Potential for additional caching strategies
- Dynamic throttle timing based on device capabilities

## Related Files
- `performanceMonitor.js`
- Game core files (coordinate and trajectory calculations)